requirements.txt
python-binance==1.0.13
requests==2.28.1
python-dotenv==0.20.0